requirements.txt
azure-storage-blob==12.19.1
openpyxl==3.1.2
pandas==2.2.1
Pillow==10.3.0
PyMySQL==1.1.0
PyYAML==6.0.1
streamlit==1.34.0
streamlit_authenticator==0.3.1
streamlit_survey==0.1.0
